==> 12_Analisi_de_Correspondencies.R <==

################################################################################
# ----- INCORPORACIÓ DADES -----
################################################################################

dades <- readRDS("dades_finals.rds") %>% 
  na.omit()

vars_socio <- c(
  "fac_Edad_media",
  "fac_Población",
  "fac_Tamaño_medio_del_hogar",
  "fac_renda_media_uc",
  "fac_indice_gini",
  "fac_poblacio_no_espanyola"
)

noms_curts <- c(
  "fac_Edad_media_Secció d'edat mitjana" = "Edat_Mitjana",
  "fac_Edad_media_Secció envellida" = "Edat_Alta",
  "fac_Edad_media_Secció jove" = "Edat_Baixa",
  "fac_Población_Població alta" = "Pob_Alta",
  "fac_Población_Població baixa" = "Pob_Baixa",
  "fac_Población_Població mitjana-alta" = "Pob_MitjanaAlta",
  "fac_Población_Població mitjana-baixa" = "Pob_MitjanaBaixa",
  "fac_Tamaño_medio_del_hogar_Llars grans" = "Llar_Gran",
  "fac_Tamaño_medio_del_hogar_Llars mitjanes" = "Llar_Mitjana",
  "fac_Tamaño_medio_del_hogar_Llars petites" = "Llar_Petita",
  "fac_indice_gini_Alta desigualtat" = "Gini_Alt",
  "fac_indice_gini_Baixa desigualtat" = "Gini_Baix",
  "fac_indice_gini_Desigualtat mitjana" = "Gini_Mitjà",
  "fac_poblacio_no_espanyola_Alta immigració" = "Immig_Alta",
  "fac_poblacio_no_espanyola_Baixa immigració" = "Immig_Baixa",
  "fac_poblacio_no_espanyola_Mitjana immigració" = "Immig_Mitjana",
  "fac_renda_media_uc_Renda alta" = "Renda_Alta",
  "fac_renda_media_uc_Renda baixa" = "Renda_Baixa",
  "fac_renda_media_uc_Renda mitjana" = "Renda_Mitjana",
  "fac_renda_media_uc_Renda molt alta" = "Renda_MoltAlta",
  "fac_renda_media_uc_Renda molt baixa" = "Renda_MoltBaixa"
)

# noms_curts <- c(
#   "Edat_Mitjana",
#   "Edat_Alta",
#   "Edat_Baixa",
#   "Pob_Alta",
#   "Pob_Baixa",
#   "Pob_MitjanaAlta",
#   "Pob_MitjanaBaixa",
#   "Llar_Gran",
#   "Llar_Mitjana",
#   "Llar_Petita",
#   "Gini_Alt",
#   "Gini_Baix",
#   "Gini_Mitjà",
#   "Immig_Alta",
#   "Immig_Baixa",
#   "Immig_Mitjana",
#   "Renda_Alta",
#   "Renda_Baixa",
#   "Renda_Mitjana",
#   "Renda_MoltAlta",
#   "Renda_MoltBaixa"
# )

llista_eleccions <- c("2015_des", "2016_jun", "2019_abr", "2019_nov", "2023_jul")
noms_eleccions   <- c("Desembre 2015", "Juny 2016", "Abril 2019", 
                      "Novembre 2019", "Juliol 2023")

################################################################################
# ----- CA I RESULTATS -----
################################################################################

for (i in seq_along(llista_eleccions)) {
  
  eleccio_id  <- llista_eleccions[i]
  eleccio_nom <- noms_eleccions[i]
  
  message("------------------------------------------------------------")
  message("Processant elecció: ", eleccio_nom)
  message("------------------------------------------------------------")
  
  # ------------------------------
  # Filtrar dades
  # ------------------------------
  dades_e <- dades %>% 
    filter(eleccio == eleccio_id) %>%
    select(eleccio, seccio, sigles, vots, all_of(vars_socio))
  
  # ------------------------------
  # Construir matriu CA
  # ------------------------------
  matriu_ca <- dades_e %>%
    pivot_longer(cols = all_of(vars_socio),
                 names_to = "variable",
                 values_to = "categoria") %>%
    unite("var_cat", variable, categoria, sep = "_") %>%
    group_by(sigles, var_cat) %>%
    summarise(vots_total = sum(vots), .groups = "drop") %>%
    pivot_wider(names_from = var_cat,
                values_from = vots_total,
                values_fill = 0)

  colnames(matriu_ca) <- ifelse(
    colnames(matriu_ca) %in% names(noms_curts),
    noms_curts[colnames(matriu_ca)],
    colnames(matriu_ca)
  )
  
  matriu_ca <- matriu_ca %>% 
    column_to_rownames("sigles")
  
  matriu_ca <- matriu_ca[rowSums(matriu_ca) > 0, ]
  
  # ------------------------------
  # TEST χ² + BALLOONPLOT
  # ------------------------------
  dt <- as.table(as.matrix(matriu_ca))
  
  balloonplot(dt,
              main = paste("Balloonplot -", eleccio_nom),
              xlab = "",
              ylab = "",
              label = FALSE,
              show.margins = FALSE)
  
  chisq <- chisq.test(matriu_ca)
  print(chisq)
  
  # ------------------------------
  # CA
  # ------------------------------
  
  res.ca <- CA(matriu_ca, graph = FALSE)
  
  print(
    fviz_ca_biplot(
      res.ca,
      repel = TRUE,
      label = "all",
      col.row = "black",
      col.col = "blue",
      title = paste("CA - Partits i factors socioeconòmics (", eleccio_nom, ")", 
                    sep = "")
    )
  )
  
  print(
    fviz_screeplot(
      res.ca,
      addlabels = TRUE,
      ylim = c(0, 85),
      title = paste("Screeplot -", eleccio_nom)
    )
  )
  
  print(
    fviz_ca_row(
      res.ca,
      col.row = "cos2",
      repel = TRUE,
      title = paste("Qualitat representació partits -", eleccio_nom)
    ) +
      scale_color_gradientn(
        colours = c("#00AFBB", "#E7B800", "#FC4E07"), 
        limits = c(0, 1),
        values = c(0, 0.5, 1)
      )
  )
  
  rows_contrib <- res.ca$row$contrib[, 1:2] %>% 
    as.data.frame() %>%
    setNames(c("Dim.1", "Dim.2")) %>%
    tibble::rownames_to_column("nom") %>%
    pivot_longer(
      cols = c(Dim.1, Dim.2),
      names_to = "Dimensio",
      values_to = "Contribucio"
    )
  
  print(
    ggplot(rows_contrib, aes(x = reorder(nom, Contribucio), y = Contribucio, fill = Dimensio)) +
      geom_col(position = "dodge") +
      coord_flip() +
      scale_fill_manual(values = c("Dim.1" = "#00AFBB", "Dim.2" = "#FC4E07")) +
      labs(
        title = paste("Contribució dels partits \nEleccions", eleccio_nom),
        x = "",
        y = "Contribució (%)"
      ) +
      theme_minimal(base_size = 13)
  )
  
  print(
    fviz_ca_col(
      res.ca,
      col.col = "cos2",
      repel = TRUE,
      title = paste("Qualitat representació variables -", eleccio_nom)
    ) +
      scale_color_gradientn(
        colours = c("#00AFBB", "#E7B800", "#FC4E07"), 
        limits = c(0, 1),
        values = c(0, 0.5, 1)
      )
  )
  
  cols_contrib <- res.ca$col$contrib[, 1:2] %>% 
    as.data.frame() %>%
    setNames(c("Dim.1", "Dim.2")) %>%
    tibble::rownames_to_column("nom") %>%
    pivot_longer(
      cols = c(Dim.1, Dim.2),
      names_to = "Dimensio",
      values_to = "Contribucio"
    )
  
  print(
    ggplot(cols_contrib, aes(x = reorder(nom, Contribucio), y = Contribucio, fill = Dimensio)) +
      geom_col(position = "dodge") +
      coord_flip() +
      scale_fill_manual(values = c("Dim.1" = "#00AFBB", "Dim.2" = "#FC4E07")) +
      labs(
        title = paste("Contribució de les variables \nEleccions", eleccio_nom),
        x = "",
        y = "Contribució (%)"
      ) +
      theme_minimal(base_size = 13)
  )
  
}
